Create factory.
